refactor: 🛠️ Restructure error handling and broker traits for improved clarity and organization
