+++ rust/src/errors.rs
@@ -4,7 +4,6 @@
 //! serialization errors.
 
 mod ack;
-mod broker;
 mod decode;
 mod encode;
 mod r#pub;
@@ -12,9 +11,9 @@ mod sub;
 mod unsub;
 
 pub use self::ack::AckError;
-pub use self::broker::BrokerError;
 pub use self::decode::DecodeError;
 pub use self::encode::EncodeError;
 pub use self::r#pub::PubError;
 pub use self::sub::SubError;
 pub use self::unsub::UnSubError;
+pub use crate::brokers::errors::BrokerError;
